AG-5670 Join duplicate bar-like series code into AbstractBarSeries
